+++ backend/de.metas.adempiere.adempiere/base/src/main/java-legacy/org/adempiere/model/ZoomInfoFactory.java
@@ -13,44 +13,23 @@
  *****************************************************************************/
 package org.adempiere.model;
 
-import java.time.Duration;
 import java.util.ArrayList;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Properties;
-import java.util.concurrent.TimeUnit;
-import java.util.function.IntSupplier;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.stream.Stream;
 
 import javax.annotation.Nullable;
 
 import org.adempiere.ad.element.api.AdWindowId;
-import org.adempiere.ad.table.api.IADTableDAO;
 import org.adempiere.exceptions.AdempiereException;
-import org.adempiere.util.lang.IPair;
-import org.adempiere.util.lang.ImmutablePair;
-import org.compiere.model.I_AD_Column;
-import org.compiere.model.MQuery;
-import org.compiere.model.PO;
-import org.compiere.model.POInfo;
-import org.compiere.util.DisplayType;
-import org.compiere.util.Env;
-import org.compiere.util.Evaluatee;
-import org.compiere.util.Evaluatees;
 import org.slf4j.Logger;
 
-import com.google.common.base.MoreObjects;
 import com.google.common.base.Stopwatch;
 import com.google.common.collect.ImmutableList;
 
-import de.metas.i18n.ITranslatableString;
-import de.metas.i18n.TranslatableStrings;
 import de.metas.logging.LogManager;
 import de.metas.security.IUserRolePermissions;
-import de.metas.util.Check;
-import de.metas.util.Services;
-import lombok.Builder;
-import lombok.Getter;
+import de.metas.util.lang.Priority;
 import lombok.NonNull;
 
 /**
@@ -65,318 +44,6 @@ public class ZoomInfoFactory
 	}
 
 	private static final transient ZoomInfoFactory instance = new ZoomInfoFactory();
-
-	public interface IZoomSource
-	{
-		Properties getCtx();
-
-		Evaluatee createEvaluationContext();
-
-		String getTrxName();
-
-		AdWindowId getAD_Window_ID();
-
-		String getTableName();
-
-		int getAD_Table_ID();
-
-		/**
-		 * The name of the column from which the zoom originates.<br>
-		 * In other words, the name of the column for which the system tries to load {@link IZoomProvider}s.
-		 *
-		 * Example: when looking for zoom-targets from a {@code C_Order} window, then this is {@code C_Order_ID}.
-		 *
-		 * @return The name of the single ID column which has {@link I_AD_Column#COLUMN_IsGenericZoomOrigin} {@code ='Y'}.
-		 *         May also return {@code null}. In this case, expect no exception, but also no zoom providers to be created.
-		 */
-		String getKeyColumnNameOrNull();
-
-		/**
-		 * @return {@code true} if the zoom source shall be considered by {@link GenericZoomProvider}.
-		 *         We have dedicated flag because the generic zoom provide can be very nice<br>
-		 *         (no need to set up relation types) but also very performance intensive.
-		 */
-		boolean isGenericZoomOrigin();
-
-		int getRecord_ID();
-
-		default boolean isSOTrx()
-		{
-			return Env.isSOTrx(getCtx());
-		}
-
-		boolean hasField(String columnName);
-
-		Object getFieldValue(String columnName);
-
-		default boolean getFieldValueAsBoolean(final String columnName)
-		{
-			return DisplayType.toBoolean(getFieldValue(columnName));
-		}
-	}
-
-	/**
-	 * Note that webui records own source implementation.
-	 */
-	public static final class POZoomSource implements IZoomSource
-	{
-		public static POZoomSource of(final PO po, final AdWindowId adWindowId)
-		{
-			return new POZoomSource(po, adWindowId);
-		}
-
-		public static POZoomSource of(final PO po)
-		{
-			final AdWindowId adWindowId = null;
-			return new POZoomSource(po, adWindowId);
-		}
-
-		private final PO po;
-		private final AdWindowId adWindowId;
-		private final String keyColumnName;
-
-		@Getter
-		private final boolean genericZoomOrigin;
-
-		private POZoomSource(@NonNull final PO po, final AdWindowId adWindowId)
-		{
-			this.po = po;
-			this.adWindowId = adWindowId;
-
-			final IPair<String, Boolean> pair = extractKeyColumnNameOrNull(po);
-
-			keyColumnName = pair.getLeft();
-			genericZoomOrigin = pair.getRight();
-		}
-
-		/**
-		 * @return the name of a key column that is also flagged as GenericZoomOrigin and {@code true},if there is exactly one such column.<br>
-		 *         Otherwise it returns {@code null} and {@code false}.
-		 */
-		private static IPair<String, Boolean> extractKeyColumnNameOrNull(@NonNull final PO po)
-		{
-			final String[] keyColumnNamesArr = po.get_KeyColumns();
-			if (keyColumnNamesArr == null)
-			{
-				return null;
-			}
-
-			final IADTableDAO adTableDAO = Services.get(IADTableDAO.class);
-
-			final ArrayList<String> eligibleKeyColumnNames = new ArrayList<>();
-			for (String element : keyColumnNamesArr)
-			{
-				final I_AD_Column column = adTableDAO.retrieveColumn(po.get_TableName(), element);
-				if (column.isGenericZoomOrigin())
-				{
-					eligibleKeyColumnNames.add(element);
-				}
-			}
-
-			if (eligibleKeyColumnNames.size() != 1)
-			{
-				return ImmutablePair.of(null, Boolean.FALSE);
-			}
-
-			return ImmutablePair.of(eligibleKeyColumnNames.get(0), Boolean.TRUE);
-		}
-
-		@Override
-		public String toString()
-		{
-			return MoreObjects.toStringHelper(this)
-					.add("po", po)
-					.add("AD_Window_ID", adWindowId)
-					.toString();
-		}
-
-		public PO getPO()
-		{
-			return po;
-		}
-
-		@Override
-		public AdWindowId getAD_Window_ID()
-		{
-			return adWindowId;
-		}
-
-		@Override
-		public String getTableName()
-		{
-			return po.get_TableName();
-		}
-
-		@Override
-		public int getAD_Table_ID()
-		{
-			return po.get_Table_ID();
-		}
-
-		@Override
-		public String getKeyColumnNameOrNull()
-		{
-			return keyColumnName;
-		}
-
-		@Override
-		public int getRecord_ID()
-		{
-			return po.get_ID();
-		}
-
-		@Override
-		public Properties getCtx()
-		{
-			return po.getCtx();
-		}
-
-		@Override
-		public String getTrxName()
-		{
-			return po.get_TrxName();
-		}
-
-		@Override
-		public Evaluatee createEvaluationContext()
-		{
-			final Properties privateCtx = Env.deriveCtx(getCtx());
-
-			final PO po = getPO();
-			final POInfo poInfo = po.getPOInfo();
-			for (int i = 0; i < poInfo.getColumnCount(); i++)
-			{
-				final Object val;
-				final int dispType = poInfo.getColumnDisplayType(i);
-				if (DisplayType.isID(dispType))
-				{
-					// make sure we get a 0 instead of a null for foreign keys
-					val = po.get_ValueAsInt(i);
-				}
-				else
-				{
-					val = po.get_Value(i);
-				}
-
-				if (val == null)
-				{
-					continue;
-				}
-
-				if (val instanceof Integer)
-				{
-					Env.setContext(privateCtx, "#" + po.get_ColumnName(i), (Integer)val);
-				}
-				else if (val instanceof String)
-				{
-					Env.setContext(privateCtx, "#" + po.get_ColumnName(i), (String)val);
-				}
-			}
-
-			return Evaluatees.ofCtx(privateCtx, Env.WINDOW_None, false);
-		}
-
-		@Override
-		public boolean hasField(final String columnName)
-		{
-			return po.getPOInfo().hasColumnName(columnName);
-		}
-
-		@Override
-		public Object getFieldValue(final String columnName)
-		{
-			return po.get_Value(columnName);
-		}
-	}
-
-	public static final class ZoomInfo
-	{
-		private final String zoomInfoId;
-		private final String _internalName;
-		private final ITranslatableString destinationDisplay;
-		private final MQuery query;
-		private final AdWindowId windowId;
-
-		private final IntSupplier recordsCountSupplier;
-
-		@Builder
-		private ZoomInfo(
-				@NonNull final String zoomInfoId,
-				@NonNull final String internalName,
-				@NonNull final AdWindowId windowId,
-				@NonNull final MQuery query,
-				@NonNull final ITranslatableString destinationDisplay,
-				@NonNull final IntSupplier recordsCountSupplier)
-		{
-			this.zoomInfoId = Check.assumeNotEmpty(zoomInfoId, "zoomInfoId is not empty");
-			this._internalName = Check.assumeNotEmpty(internalName, "internalName is not empty");
-
-			this.windowId = windowId;
-
-			this.query = query;
-			this.destinationDisplay = destinationDisplay;
-
-			this.recordsCountSupplier = recordsCountSupplier;
-		}
-
-		@Override
-		public String toString()
-		{
-			return MoreObjects.toStringHelper(this)
-					.add("zoomInfoId", zoomInfoId)
-					.add("internalName", _internalName)
-					.add("display", destinationDisplay)
-					.add("AD_Window_ID", windowId)
-					.add("RecordCount", query.getRecordCount())
-					.toString();
-		}
-
-		public String getId()
-		{
-			return zoomInfoId;
-		}
-
-		public String getInternalName()
-		{
-			return _internalName;
-		}
-
-		public ITranslatableString getLabel()
-		{
-			final ITranslatableString postfix = TranslatableStrings.constant(" (#" + getRecordCount() + ")");
-			return TranslatableStrings.join("", destinationDisplay, postfix);
-		}
-
-		public int getRecordCount()
-		{
-			return query.getRecordCount();
-		}
-
-		@Nullable
-		public Duration getRecordCountDuration()
-		{
-			return query.getRecordCountDuration();
-		}
-
-		public void updateRecordsCount()
-		{
-			final Stopwatch stopwatch = Stopwatch.createStarted();
-			final int recordsCount = recordsCountSupplier.getAsInt();
-			final Duration recordsCountDuration = Duration.ofNanos(stopwatch.stop().elapsed(TimeUnit.NANOSECONDS));
-			query.setRecordCount(recordsCount, recordsCountDuration);
-			logger.debug("Updated records count for {} in {}", this, stopwatch);
-		}
-
-		public AdWindowId getAdWindowId()
-		{
-			return windowId;
-		}
-
-		public MQuery getQuery()
-		{
-			return query;
-		}
-	}
 
 	private static final Logger logger = LogManager.getLogger(ZoomInfoFactory.class);
 
@@ -407,29 +74,30 @@ public class ZoomInfoFactory
 		return streamZoomInfos(zoomOrigin, onlyTargetWindowId, rolePermissions);
 	}
 
-	public Stream<ZoomInfo> streamZoomInfos(
-			@NonNull final IZoomSource zoomOrigin,
+	private Stream<ZoomInfo> streamZoomInfos(
+			@NonNull final IZoomSource zoomSource,
 			@Nullable final AdWindowId onlyTargetWindowId,
 			@NonNull final IUserRolePermissions rolePermissions)
 	{
-		logger.debug("source={}", zoomOrigin);
+		logger.debug("source={}", zoomSource);
 
-		final ImmutableList<ZoomInfo> zoomInfoCandidates = getZoomInfoCandidates(zoomOrigin, onlyTargetWindowId);
+		final ImmutableList<ZoomInfo> zoomInfoCandidates = getZoomInfoCandidates(zoomSource, onlyTargetWindowId, rolePermissions);
 
-		final HashSet<AdWindowId> alreadySeenWindowIds = new HashSet<>();
+		final ConcurrentHashMap<AdWindowId, Priority> alreadySeenWindowIds = new ConcurrentHashMap<>();
 
 		return zoomInfoCandidates.stream()
 				.sequential()
-				.filter(zoomInfo -> isEligible(zoomInfo, rolePermissions, alreadySeenWindowIds));
+				.filter(zoomInfo -> isEligible(zoomInfo, alreadySeenWindowIds));
 	}
 
 	private ImmutableList<ZoomInfo> getZoomInfoCandidates(
-			@NonNull final IZoomSource zoomOrigin,
-			@Nullable final AdWindowId onlyTargetWindowId)
+			@NonNull final IZoomSource zoomSource,
+			@Nullable final AdWindowId onlyTargetWindowId,
+			@NonNull final IUserRolePermissions rolePermissions)
 	{
 		final Stopwatch stopwatch = Stopwatch.createStarted();
 
-		final String tableName = zoomOrigin.getTableName();
+		final String tableName = zoomSource.getTableName();
 		final List<IZoomProvider> zoomProviders = retrieveZoomProviders(tableName);
 
 		final ImmutableList.Builder<ZoomInfo> zoomInfoCandidates = ImmutableList.builder();
@@ -437,55 +105,58 @@ public class ZoomInfoFactory
 		{
 			try
 			{
-				final List<ZoomInfo> zoomInfos = zoomProvider.retrieveZoomInfos(zoomOrigin, onlyTargetWindowId);
+				final List<ZoomInfo> zoomInfos = zoomProvider.retrieveZoomInfos(zoomSource, onlyTargetWindowId);
 				for (final ZoomInfo zoomInfo : zoomInfos)
 				{
 					// If not our target window ID, skip it
 					// This shall not happen because we asked the zoomProvider to return only those for our target window,
 					// but if is happening (because of a bug zoom provider) we shall not be so fragile.
-					if (onlyTargetWindowId != null && !AdWindowId.equals(onlyTargetWindowId, zoomInfo.getAdWindowId()))
+					if (onlyTargetWindowId != null
+							&& !AdWindowId.equals(onlyTargetWindowId, zoomInfo.getAdWindowId()))
 					{
 						new AdempiereException("Got a ZoomInfo which is not for our target window. Skipping it."
 								+ "\n zoomInfo: " + zoomInfo
 								+ "\n zoomProvider: " + zoomProvider
 								+ "\n targetAD_Window_ID: " + onlyTargetWindowId
-								+ "\n source: " + zoomOrigin)
+								+ "\n source: " + zoomSource)
 										.throwIfDeveloperModeOrLogWarningElse(logger);
 						continue;
 					}
 
+					//
+					// Filter out those windows on given user does not have permissions
+					if (!rolePermissions.checkWindowPermission(zoomInfo.getAdWindowId()).hasReadAccess())
+					{
+						continue;
+					}
+
+					//
+					// Collect eligible zoom info candidate
 					zoomInfoCandidates.add(zoomInfo);
 				}
-				zoomInfoCandidates.addAll(zoomInfos);
 			}
 			catch (final Exception ex)
 			{
-				logger.warn("Failed retrieving zoom infos from {} for {}. Skipped.", zoomProvider, zoomOrigin, ex);
+				logger.warn("Failed retrieving zoom infos from {} for {}. Skipped.", zoomProvider, zoomSource, ex);
 			}
 		}
 
 		stopwatch.stop();
-		logger.debug("Fetched zoom candidates for source={} in {}", zoomOrigin, stopwatch);
+		logger.debug("Fetched zoom candidates for source={} in {}", zoomSource, stopwatch);
 
 		return zoomInfoCandidates.build();
 	}
 
-	private boolean isEligible(
+	private static boolean isEligible(
 			@NonNull final ZoomInfo zoomInfo,
-			@NonNull final IUserRolePermissions rolePermissions,
-			@NonNull final HashSet<AdWindowId> alreadySeenWindowIds)
+			@NonNull final ConcurrentHashMap<AdWindowId, Priority> alreadySeenWindowIds)
 	{
-		//
-		// Filter out those windows on given user does not have permissions
-		if (!rolePermissions.checkWindowPermission(zoomInfo.getAdWindowId()).hasReadAccess())
-		{
-			return false;
-		}
-
 		//
 		// Only consider a window already seen if it actually has record count > 0 (task #1062)
 		final AdWindowId adWindowId = zoomInfo.getAdWindowId();
-		if (alreadySeenWindowIds.contains(adWindowId))
+		final Priority alreadySeenZoomInfoPriority = alreadySeenWindowIds.get(adWindowId);
+		if (alreadySeenZoomInfoPriority != null
+				&& alreadySeenZoomInfoPriority.isHigherThan(zoomInfo.getPriority()))
 		{
 			logger.debug("Skipping zoomInfo {} because there is already one for destination '{}'", zoomInfo, adWindowId);
 			return false;
@@ -506,7 +177,7 @@ public class ZoomInfoFactory
 		//
 		// We got a valid zoom info
 		// => accept it
-		alreadySeenWindowIds.add(adWindowId);
+		alreadySeenWindowIds.put(adWindowId, zoomInfo.getPriority());
 		return true;
 	}
 
@@ -531,12 +202,11 @@ public class ZoomInfoFactory
 				.orElseThrow(() -> new AdempiereException("No zoomInfo found for source=" + zoomSource + ", targetWindowId=" + targetWindowId));
 	}
 
-	private List<IZoomProvider> retrieveZoomProviders(final String tableName)
+	private List<IZoomProvider> retrieveZoomProviders(@NonNull final String tableName)
 	{
-		final List<IZoomProvider> zoomProviders = new ArrayList<>();
+		// NOTE: Zoom providers order IS NOT important because each provider creates ZoomInfo with a priority.
 
-		// NOTE: Zoom providers order matter because in case it finds some duplicates (i.e. same window),
-		// it will pick only the first one (i.e. the one from the first provider).
+		final ArrayList<IZoomProvider> zoomProviders = new ArrayList<>();
 		zoomProviders.addAll(RelationTypeZoomProvidersFactory.instance.getZoomProvidersByZoomOriginTableName(tableName));
 		zoomProviders.add(GenericZoomProvider.instance);
 		if (factAcctZoomProviderEnabled)
