+++ backend/metasfresh-webui-api/src/main/java/de/metas/ui/web/window/model/DocumentReference.java
@@ -7,7 +7,7 @@ import javax.annotation.Nullable;
 import de.metas.i18n.ITranslatableString;
 import de.metas.ui.web.document.filter.DocumentFilter;
 import de.metas.ui.web.window.datatypes.WindowId;
-
+import de.metas.util.lang.Priority;
 import lombok.Builder;
 import lombok.NonNull;
 import lombok.Value;
@@ -25,11 +25,11 @@ import lombok.Value;
  *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public
- * License along with this program.  If not, see
+ * License along with this program. If not, see
  * <http://www.gnu.org/licenses/gpl-2.0.html>.
  * #L%
  */
@@ -41,16 +41,18 @@ public final class DocumentReference
 	private final String internalName;
 	private final ITranslatableString caption;
 	private final WindowId windowId;
+	private final Priority priority;
 	private final int documentsCount;
 	private final DocumentFilter filter;
 	private final Duration loadDuration;
-	
+
 	@Builder
 	private DocumentReference(
 			@NonNull final String id,
 			@Nullable final String internalName,
 			@NonNull final ITranslatableString caption,
 			@NonNull final WindowId windowId,
+			@NonNull final Priority priority,
 			final int documentsCount,
 			@NonNull final DocumentFilter filter,
 			@Nullable final Duration loadDuration)
@@ -59,6 +61,7 @@ public final class DocumentReference
 		this.internalName = internalName;
 		this.caption = caption;
 		this.windowId = windowId;
+		this.priority = priority;
 		this.documentsCount = documentsCount;
 		this.filter = filter;
 		this.loadDuration = loadDuration;
